refactor(types): move to root
