+++ tsconfig.json
@@ -30,7 +30,7 @@
     /* Generates corresponding '.map' file. */
     // "outFile": "./dist/index.js",                       /* Concatenate and emit output to single file. */
     "outDir": "./dist" /* Redirect output structure to the directory. */,
-    "rootDir": "plugins" /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */,
+    "rootDir": "." /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */,
     "composite": false, /* Enable project compilation */
     "removeComments": true /* Do not emit comments to output. */,
     // "noEmit": true,                        /* Do not emit outputs. */
