+++ types/scene.ts
